l2-gemini/pyproject.toml
[project]
name = "l2-gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.25.0",
]
